refactor(registry): rename IaCM toolset to infrastructure; fold CD infrastructure into environments
Consolidate infrastructure-related capabilities under clearer names, with no
breaking changes — every old name still resolves via the alias layer.

- Rename the `iacm` toolset to `infrastructure` (old `iacm` name kept as a toolset alias).
- Rename the 7 IaCM resource types `iacm_*` -> `infrastructure_*`, each keeping its
  old name as a resourceType alias and gaining an `iacm` search alias.
- Move the CD `infrastructure` resource into the `environments` toolset and rename it
  to `infrastructure_definition`, keeping `infrastructure` as an alias.
- Delete the standalone CD infrastructure toolset; update ToolsetName union and tests.

Routing, scope, request bodies, extractors, and deep links are unchanged; only the
canonical names differ, so existing callers using the old names keep working.


AI-Session-Id: 6f92026d-176b-4d67-8b5d-16201370f3cb
AI-Tool: claude-code
AI-Model: unknown
